Add thread-safety and Ruby 2.7 support
Why:

If we simply use a class variable, a second thread can replace
I18n.config.globals after the first one has been set.

This is how it happens:
1. Thread A: I18n.config.globals = {foo: 'bar'}
2. Thread B: I18n.config.globals = {foo: 'baz'}
3. Thread A now calls I18n.t('x') to interpolate, getting B's globals

Below is an example of how to test this. Assume we load certain globals
from the current user.

1. Set up a delay for a user just after the globals are assigned:

before_action {
  I18n.config.globals = I18n.config.globals.merge(current_user.global_locale_variables)
  sleep(10) if current_user.email == 'user1@gmail.com'
}

2. Load the page for this user
3. Immediately load the page for a second user, during the 10s wait
4. Now the first user will have the second user's globals

For this fork I have reset Master to an earlier Master:
113110cd91fa39a278adbe000c354b09a2b2a06b

The reason is that the published Gem uses the earlier code, we don't
need the much more complex feature that was added after that, and
the simpler i18n-globals has been working for us in Production.

I have also merged in this small change to support Ruby 2.7 and above:
https://github.com/nusnick/i18n-globals/commit/8e9bfd57ac2cf492a2d17b9d81250f2567ce7ca6